Rewritten and revised Intro and SoA
They have to be rewritten completely. The SoA is the usual this guys do this, some other guys do that. We should present an argument for our paper, along the lines of

* Prediction is usually based on time series forecasting (refs) or, in some cases, in x or y from blogs (refs).
* however, predicting the reaction to new products is hard for two reasons: no historial data (refs) and difficult processing of available variables: categorical, non-relevant, whatever (refs).
* Even so, these guys (refs) have managed to do this and that, but they are black box (refs) <- there is no refs for this last one.

So, our method will first process data to make it more manageable and extract relevant variables and then use the result to predict sales via a model that can also be used as a decision making support system, explaining the reasons why...

Please @amorag check this out and rewrite. When you're done, send me for revision