Mainly comments on the paper
And I think the UMatrix should be moved to Section 3, giving it some more weight. Besides, UMatrix is done for exploratory data analysis and to examine the problem, so it corresponds to "problem description". It is really _not_ part of data processing, because you don't extract any conclusion that excludes or includes variables (I think). Please @amorag check this out.