Revising the state of the art.
it still has many suggestions and things to address. First thing, it lacks a narrative. The state of the art should be possibly resumed in a sentence, such as
"So far, most forecasting techniques have been based on previous sales. Very few have used, like we do in this paper, pre-sales data. So this is what we are doing here and we will improve methods"

Please @fergunet @amorag @pacastillo take a long hard look at it, and as done before address suggestions and issues and erase them when done.