Working on the explanation of the trees
This is one of the main results of our paper and it was dispatched with a couple of lookie here, this graphic and some paragraphs. What we have to do in this paper is to actually look at the factors that make a book sell, and that is contained within those trees so I have added a good amount of explanation to them.

Refs #39 #25