Waiting at the airport
And revising introduction as instructed by #45. Maybe @pacastillo or @fergunet could have a look. I have the impression that it is excessively iterative and that it does not follow a clear line of reasoning, not to mention that it does not hightlight our actual contribution. I will continue with this anyway.